장소 검색 tab 수정
