+++ app/src/main/java/com/yeonae/chamelezone/view/search/SearchActivity.kt
@@ -1,14 +1,12 @@
 package com.yeonae.chamelezone.view.search
 
 import android.os.Bundle
-import android.util.Log
 import android.view.inputmethod.EditorInfo
 import androidx.appcompat.app.AppCompatActivity
-import androidx.fragment.app.FragmentStatePagerAdapter
-import com.google.android.libraries.places.api.model.Place
+import androidx.fragment.app.Fragment
 import com.google.android.material.tabs.TabLayout
 import com.yeonae.chamelezone.R
-import com.yeonae.chamelezone.view.search.adapter.SearchTabAdapter
+import com.yeonae.chamelezone.adapter.PagerAdapter
 import kotlinx.android.synthetic.main.activity_search.*
 
 class SearchActivity : AppCompatActivity(), KeywordTabFragment.OnKeywordSelectedListener {
@@ -17,20 +15,28 @@ class SearchActivity : AppCompatActivity(), KeywordTabFragment.OnKeywordSelected
         edt_search.setSelection(edt_search.length())
     }
 
+    private val tabList by lazy { listOf("장소명", "지역명", "키워드명") }
+
+    private val tabPagerAdapter = object : PagerAdapter(supportFragmentManager, tabList) {
+        override fun getItem(position: Int): Fragment {
+            return when (position) {
+                0 -> PlaceNameTabFragment.newInstance()
+                1 -> AddressTabFragment.newInstance()
+                else -> KeywordTabFragment.newInstance()
+            }
+        }
+    }
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_search)
 
-        search_tab.addTab(search_tab.newTab().setText("장소명"))
-        search_tab.addTab(search_tab.newTab().setText("지역명"))
-        search_tab.addTab(search_tab.newTab().setText("키워드"))
-
-        val searchTabAdapter = SearchTabAdapter(supportFragmentManager)
-        search_view_pager.adapter = searchTabAdapter
+        search_tab.setupWithViewPager(search_view_pager)
+        search_view_pager.adapter = tabPagerAdapter
         search_view_pager.addOnPageChangeListener(TabLayout.TabLayoutOnPageChangeListener(search_tab))
-        
+
         edt_search.setOnEditorActionListener { textView, i, keyEvent ->
-            if(i == EditorInfo.IME_ACTION_DONE || i == EditorInfo.IME_ACTION_NEXT || i == EditorInfo.IME_ACTION_SEARCH || i == EditorInfo.IME_ACTION_GO){
+            if (i == EditorInfo.IME_ACTION_DONE || i == EditorInfo.IME_ACTION_NEXT || i == EditorInfo.IME_ACTION_SEARCH || i == EditorInfo.IME_ACTION_GO) {
                 supportFragmentManager.fragments.forEach {
                     when (it) {
                         is PlaceNameTabFragment -> it.searchByName("${edt_search.text}")
@@ -41,21 +47,6 @@ class SearchActivity : AppCompatActivity(), KeywordTabFragment.OnKeywordSelected
             }
             true
         }
-
-        search_tab.addOnTabSelectedListener(object : TabLayout.OnTabSelectedListener {
-            override fun onTabReselected(tab: TabLayout.Tab?) {
-            }
-
-            override fun onTabUnselected(tab: TabLayout.Tab?) {
-            }
-
-            override fun onTabSelected(tab: TabLayout.Tab?) {
-                if (tab != null) {
-                    search_view_pager.currentItem = tab.position
-                }
-            }
-
-        })
 
         btn_back.setOnClickListener {
             finish()
